// client/src/interfaces/TicketData.tsx
import { UserData } from './UserData';

export interface TicketData {
  id: number | null;
  name: string | null;
  description: string | null;
  status: string | null;
  assignedUserId: number | null;
  assignedUser: UserData | null;
  priority?: number; // Optional priority property for sorting tickets
}